+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
 
 ### 8.5
-Add function decorator `ribosome.function`
+* Add function decorator `ribosome.function`
+* Add transient `ScratchMachine` that handles mappings in a scratch buffer